+++ waypoints/src/main/kotlin/de/md5lukas/waypoints/config/general/TeleportConfiguration.kt
@@ -4,6 +4,7 @@ import de.md5lukas.konfig.ConfigPath
 import de.md5lukas.konfig.Configurable
 import de.md5lukas.konfig.TypeAdapter
 import de.md5lukas.konfig.UseAdapter
+import de.md5lukas.waypoints.api.Type
 import de.md5lukas.waypoints.util.Expression
 import de.md5lukas.waypoints.util.MathParser
 import java.time.Duration
@@ -31,6 +32,9 @@ class TeleportConfiguration {
 class TypedTeleportConfiguration {
 
   var cooldown: Duration = Duration.ZERO
+    private set
+
+  var alsoApplyCooldownTo: List<Type> = emptyList()
     private set
 
   var mustVisit: Boolean? = false